propose better nested list structure
Sorry Jilayne for stepping on your file, but I thought I would propose a fix since I raised this issue on today's call. 

Many paragraphs in the CC-BY licenses have paragraph numbers (1, 2, 3) and subparagraph letters (a, b, c). I propose nesting the "list" tag for subparagraphs within the "li" tag for main paragraphs, so that the structure reflects that 7(a) is always child of (7) and that (7) is always a sibling of (8).

@myndzi does that work for you?