removed 'trend' typo
This is a weird typo. The license page is no longer available at http://bits.netizen.com.au/licenses/NOSL/nosl.txt

_BUT_ archive.org has several archived versions of that page. Some of them have a link called "trend" at this location, which is where we must have picked it up, but the older pages do not.

Link present:
* https://web.archive.org/web/20120810055740/http://bits.netizen.com.au/licenses/NOSL/nosl.txt
* https://web.archive.org/web/20131001130911/http://bits.netizen.com.au/licenses/NOSL/nosl.txt
* https://web.archive.org/web/20131204004715/http://bits.netizen.com.au/licenses/NOSL/nosl.txt

Link not present:
 * https://web.archive.org/web/20090321184338/http://bits.netizen.com.au/licenses/NOSL/nosl.txt
 * https://web.archive.org/web/20101223005600/http://bits.netizen.com.au/licenses/NOSL/nosl.txt
 * https://web.archive.org/web/20110709074143/http://bits.netizen.com.au/licenses/NOSL/nosl.txt
 * https://web.archive.org/web/20111211060114/http://bits.netizen.com.au/licenses/NOSL/nosl.txt


My guess is that this was a spam link added by someone who got control over the website (legitimately or by breaking into it). I've removed the word "trend" from the license for now, but note that we may need to add it back and make it optional if others pulled it into their licenses in the wild the same way we got it into our license text, if we find people using the license with the extra word out in the wild.