src/*: Add <urls> to licenses that were missing them
In the GPL-3.0-with-autoconf-exception case, this is just fixing the
XML markup for an existing URL.

For the other licenses and exceptions, I've found URLs for an early
occurence in the original project.  Ideally these URLs are for a page
about the license/exception by the license/exception author, but where
that doesn't exist a link to the first known consumer seems better
than nothing.

This fixes everything turned up by:

  $ grep -L '<url>' $(git ls-tree -r --name-only HEAD | grep .xml)

except for Zimbra-1.3.  The FSF has an entry for that license [1], but
they link to [2], which Zimbra redirects to the 1.4 version:

  $ curl -I http://www.zimbra.com/license/zimbra-public-license-1-3.html
  HTTP/1.1 301 Moved Permanently
  Content-Type: text/html; charset=iso-8859-1
  Date: Fri, 09 Jun 2017 16:20:15 GMT
  Location: http://www.zimbra.com/license/zimbra-public-license-1-4.html
  ...

Zimbra's legal page only lists the 1.4 version [3], and their
robots.txt did not allow the Internet Archive to archive their site
when they had a 1.3 version [4].

[1]: https://www.gnu.org/licenses/license-list.html#Zimbra
[2]: http://www.zimbra.com/license/zimbra-public-license-1-3.html
[3]: http://www.zimbra.com/legal/licensing/
[4]: https://web.archive.org/web/*/http://zimbra.com