src/AFL-*: Use Internet Archive link
That host no longer exists:

  $ ping opensource.linux-mirror.org
  ping: unknown host opensource.linux-mirror.org

It would be better to link to opensource.org, because we're claiming
that these are all OSI-approved licenses.  But the OSI is not listing
them on its current page (probably an oversight [1]).  So instead I'm
linking to Internet Archive pages showing the old licenses on the OSI
site.

The OSI had also approved the AFL-1.0 [2], but we don't have XML for
that yet (and we may never have assigned an SPDX identifier to it).

I only have the .txt link for the AFL-2.0, because opensource.org was
still listing AFL-1.2 on 2006-09-24 [3] and had switched to the
AFL-3.0 by the next-archived academic.php in 2007-07-14 [4].

I've kept the dead links based on:

On Thu, Sep 14, 2017 at 06:05:39PM +0000, Bradlee H. Edmndson wrote [5]:
> @wking can you do the same here? (i.e. update the PR to add URLs
> instead of replace.) thanks!

Although we don't have an offical policy on dead links yet [6].

[1]: https://lists.opensource.org/pipermail/license-discuss/2017-July/019922.html
     Subject: Some formerly-approved licenses not listed under opensource.org/licenses/
     Date: Fri, 14 Jul 2017 12:22:48 -0700
     Message-ID: <20170714192248.GX17902@valgrind.tremily.us>
[2]: http://wayback.archive.org/web/20020805135807/http://www.opensource.org/licenses/academic.php
[3]: http://wayback.archive.org/web/20060924132044/http://www.opensource.org/licenses/academic.php
[4]: http://wayback.archive.org/web/20070714181427/http://www.opensource.org/licenses/academic.php
[5]: https://github.com/spdx/license-list-XML/pull/419#issuecomment-329563093
[6]: https://github.com/spdx/license-list-XML/issues/433