EUDatagrid: Remove/replace curly quotes
The straight quotes were part of early drafts [1], but when the "EU
DataGrid" entry was added (between 2002-06-25 [1] and 2002-08-17 [2])
it came in with curly quotes, which remained through the first
non-draft version archived by the Internet Archive [3].  Our matching
guidelines suggest that all quotes are equivalent [4], so I've changed
to straight quotes for internal consistency.

The (duplicated) curly quotes opening the notice are in neither the
first archived non-draft version [3] nor the earlier drafts [1,2], so
I've just dropped them.

[1]: http://wayback.archive.org/web/20020625131338/http://eu-datagrid.web.cern.ch/eu-datagrid/license.html
[2]: http://wayback.archive.org/web/20020817002230/http://eu-datagrid.web.cern.ch/eu-datagrid/license.html
[3]: http://wayback.archive.org/web/20021229214435/http://eu-datagrid.web.cern.ch:80/eu-datagrid/license.html
[4]: https://spdx.org/spdx-license-list/matching-guidelines
     Guideline 5.1.3