FTL: Remove <copyright>
From [1]:

  Copyright tag: denotes an optional (as per Matching Guideline 9)
  section of the license content, with the implication that it
  contains copyright notices *that apply to the code under license*;
  this is explicitly *not* intended to include copyright notices that
  apply to the license itself

This isn't particularly clear-cut from the source, but the bump in [2]
suggests that the line is intended to apply to the license itself.
Note that [2] bumps the copyright statement in the README to:

  Copyright 2001, 2002, 2003, 2004, 2006 by David Turner, Robert
  Wilhelm, and Werner Lemberg

but that it only bumps the FTL.TXT copyright to:

  Copyright 1996-2002, 2006 by David Turner, Robert Wilhelm, and
  Werner Lemberg

which does not include 2004.

Regardless of the author intentions, the remainder of the FTL
"license" is clearly FreeType-specific, including a contact section.
So it seems very unlikely that another project would use the same
license and only replace the line formerly contained within
<copyright>.

[1]: https://wiki.spdx.org/view/Legal_Team/Templatizing/tags-matching
[2]: http://git.savannah.gnu.org/cgit/freetype/freetype2.git/commit/?id=68a15ebbe4f61999368a72a2953e257513a3cbba