ISC: make "/or" optional in "and/or"
OpenBSD uses an earlier form of this license with a bare "and".  There
has been some discussion in the past about whether, as the University
of Washington at some point claimed, the "and" form forbids
distributing modified copies (vs. having local copies and distributing
unmodified copies, but not distributing the modified copies).  For
more on the discussion, see [1,2].  However, the SPDX legal team
decided on 2015-04-30 that *it* didn't see significant difference
between the "and" and "and/or" forms [3].  This commit encodes that
decision for use in matching software.

[1]: https://www.gnu.org/licenses/license-list.html#ISC
[2]: https://lists.debian.org/debian-legal/2000/08/msg00147.html
[3]: https://lists.spdx.org/pipermail/spdx-legal/2015-April/001398.html
     Subject: New License/Exception Request
     Date: Thu Apr 30 17:56:52 UTC 2015