schema/ListedLicense: Punt license and exception text into <text>
As it stands, there's not an easy way to extract the license text from
the XML format without removing <crossRefs>, <notes>,
<standardLicenseHeader>, and other siblings.  With this change (which,
if accepted, I still have to propagate into src/) we collect it in
<text>, just like we already collect the header text in
<standardLicenseHeader>.

Also use <all> instead of <choice> for <LicenseType> children.  With
that, <crossRefs> could have occured multiple times, etc.  With this
change, it can only occur once, although the children can still appear
in any order (we'd use <sequence> if we cared about child order).