CC-BY-ND-2.0: Return to CC-BY-ND-2.0 (had CC-BY-ND-3.0 text)
The initial entry in this repo had the 2.0 text, but f12d5fea (Update
CC-BY-ND-2.0, 2016-09-15, #27) injected the 3.0 wording my mistake.
This commit returns us to the current upstream wording for the 2.0
license.

The <optional> space unambiguously allows the upstream's (accidental?)
current lack of a space there:

  $ curl -s https://creativecommons.org/licenses/by-nd/2.0/legalcode | grep 'Restrictions.'
  <p><strong>4. Restrictions.</strong>The license granted in Section 3 above...

The <optional> s in https accounts for the upstream change between [1]
and [2].

[1]: https://web.archive.org/web/20141119175856/https://creativecommons.org/licenses/by-nd/2.0/legalcode
[2]: https://web.archive.org/web/20141220194853/https://creativecommons.org/licenses/by-nd/2.0/legalcode