LPPL-1.3a: Remove percent from "pig.ins % and"
This is an artifact from unwinding upstream's [1]:

  % This work consists of the files pig.dtx and pig.ins
  % and the derived file pig.sty.

and forgetting to remove the comment character.  The typo dates back
to ad98f3bd (Automated, 2016-03-30, #114).

The typo was also present in the released v2.6:

  $ curl -s https://raw.githubusercontent.com/spdx/license-list-data/v2.6/json/details/LPPL-1.3a.json |
  >   jq -r .standardLicenseHeader | grep 'pig\.ins'
  This work consists of <<var;name=the work;original=pig.dtx and pig.ins % and the derived file pig.sty;match=.+>>.

But since this text is already inside a wildcard <alt> we can remove
the percent without needing a new <alt> to cover folks using the
typo'ed version.

[1]: https://www.latex-project.org/lppl/lppl-1-3a.txt