Added EUPL-1.2
I used the same bullet characters as the original license, which is why it differs from the characterused by the EUPL-1.1.xml file.

TODO: Do note that the fact that the link to the OSI license list points to the same URL as the EUPL-1.1. This is because the OSI website has a bug where it seems that instead of adding a new page for EUPL-1.2 they seem to have overwritten it. I will inform OSI about it, but the <crossRef> will have to be edited if/when the OSI lists the EUPL-1.2 under a more appropriate URL.

ref:
https://opensource.org/licenses/EUPL-1.1
https://web.archive.org/web/20171208111154/https://opensource.org/licenses/EUPL-1.1