GCC-exception-2.0: Match assorted upstream versions
This exception grew an upstream URL in 23b6d5a5 (src/*: Add <urls> to
licenses that were missing them, 2017-06-09, #413), but we didn't
compare the text carefully enough in that case.  This commit adds
optional and alt blocks so we match both the upstream texts and the
old SPDX text.  Checking the mailing lists, we have Jilayne reporting
uncertainty about the source used for this exception [1,2].  But
digging through GCC's version-control history turns up this initial
version [3] (which, with this commit, becomes our canonical version).
The "combine executable" form seems to be from [4].  I've also allowed
for "combined executable", which first lands in [5].  [5] also
includes the first instance of the Library wording (although I'm not
clear on why you'd need this exception if you were already using an
LGPL).

[1]: https://lists.spdx.org/pipermail/spdx/2010-November/000204.html
     Subject: newly uploaded license list
     Date: Thu Nov 18 12:38:57 MST 2010

     > Exceptions added:
     >
     > GPL v2 w/GCC Runtime Library exception - I could not find a
     > link to this.  GNU website only have v3 exception.  Please
     > check if this text is correct...

[2]: https://lists.spdx.org/pipermail/spdx-legal/2012-September/000692.html
     Subject: SPDX License List: GPL GCC, runtime, Classpath, linking exceptions
     Date: Mon Sep 24 17:28:04 UTC 2012

     > --> in this case, then, is the GPLv2 GCC Runtime library
     > exception we do have correct?  I could not find a link to it
     > (and don't know who submitted the text we have as it was at the
     > very beginning)

[3]: https://gcc.gnu.org/git/?p=gcc.git;a=blob;f=gcc/libgcc1.c;h=762f5143fc6eed57b6797c82710f3538aa52b40b;hb=cb143a3ce4fb417c68f5fa2691a1b1b1053dfba9#l10
     cb143a3ce (Initial revision, 1992-02-15)
[4]: https://gcc.gnu.org/git/?p=gcc.git;a=commitdiff;h=4c9b6e719ad35162a897aa09936acce597d0848d;hp=6e36176ee3f04651030cfc6669e99f5b19934573
     4c9b6e719 (Update to libgcc copyright exception clause, 2000-05-15)
[5]: https://gcc.gnu.org/git/?p=gcc.git;a=commitdiff;h=ac5e00d3f37fa940ed51493fa157b016a97c35ae;hp=d1f15f1bce018b6f30830858bf2a5a2f205bd8a5
     ac5e00d3f (cplus-dem.c: Add license exception to copyright notice, 2003-03-04)