gulpfile: Drop debugging code
These were committed in 378fe01b (Update all source files to the new
XML format and add in the tools from the schemadev branch,
2017-10-17), but they are redundant.  The validateall call is
redundant when you run the 'validate' task (which has its own
validateall call).  And the console.log('done') prints done, when it
would be better to have validateall print a summary.