BSD-1-Clause: Add a period before "All rights reserved"
Upstream [1] has:

  * Copyright (c) 1995, 1999
  *      Berkeley Software Design, Inc.  All rights reserved.

Which collapses the period from "Inc." and the sentence-terminating
period into a single period.  However, owners which do not end in a
period will still need a sentence-terminating period (compare with out
current BSD-3-Clause content).  This commit requires the
sentence-terminating period, regardless of whether it is also serving
as an owner-abbreviation period.

[1]: https://svnweb.freebsd.org/base/head/include/ifaddrs.h?revision=326823