.github/issue_template: Add an issue template
The old docs for this process are in [1] as well as the CONTRIBUTING
file I'm adjusting here.  This template pulls out the subset of
information required by the initial request, which should make life
more convenient for folks using GitHub issues to propose new
licenses/exceptions.

I've also tweaked two CONTRIBUTING headers, because the "general"
vs. "other" distinction didn't seem very clear to me from just the
headers, and I needed to link to the mailing-list section from the
submission procedure.

Docs for the filename are in [2].  I've used a .github subdirectory to
keep this peripheral workflow information out of the root directory.

GitHub also supports pull-request templates [3], but most of this
information will be included in the pull-request payload, and doesn't
need to be repeated in the pull-request's opening comment.  The
exception is the short motivation, but I'm comfortable assuming folks
who file pull requests know enough to provide that without prompting.

[1]: https://spdx.org/spdx-license-list/request-new-license
[2]: https://help.github.com/articles/creating-an-issue-template-for-your-repository/
[3]: https://help.github.com/articles/creating-a-pull-request-template-for-your-repository/