AGPL-3.0: Drop redundant release-date <notes>
The XSD suggests (buggily) that we only want a single <notes> element
(more details on this in 5010132f, wxWindows: Collapse to a single
<notes> element, 2017-12-27, #577).  And the information contained in
the <notes> section I'm removing is redundant, because the titleText
contains the release date.  I don't see a need to include that release
date as unstrucutured information in two places, and the license title
is clearly the more important location.