src: Move <standardLicenseHeader> inside <text> where possible
Catch up with the previous commits' schema change.

To ease review, this commit just injects the standardLicenseHeader
tags on new lines but tries not to touch the indent or other
pretty-printing of the wrapped content.

Some suggested grant text contains variable text, and the embedded
<standardLicenseHeader> elements do not distinguish between "variable
for the purpose of license matching" and "variable for the purpose of
grant matching".  For example:

  <license>
    <text>
      <p>Use this grant:</p>
      <standardLicenseHeader>
        <p>Copyright <alt name="copyright" match=".+">[year] [name]</alt>.
        All rights reserved.</p>
      </standardLicenseHeader>
    <text>
  </license>

is a bit different from:

  <license>
    <standardLicenseHeader>
      <p>Copyright <alt name="copyright" match=".+">[year] [name]</alt>.
      All rights reserved.</p>
    </standardLicenseHeader>
    <text>
      <p>Use this grant:</p>
      <p>Copyright [year] [name].
      All rights reserved.</p>
    <text>
  </license>

because the former allows for variation in both the license and the
grant, while the later allows variation in the grant but not in the
license.  In cases where we feel that distinction is significant,
we'll want to go back to sibling standardLicenseHeader elements.

This commit fixes a number of issues:

* ECL-1.0: Fix a missing paragraph in the old grant text, which had
  lead to the copyright holders running into the next sentence.  With
  this commit, each sentence in the grant text is its own paragraph.
  The same issue exists in the ECL-2.0, but since both our previous
  license text and the upstream text seem to have the issue, I've left
  it alone for this commit.

* GFDL-*: The replacable invariant and cover-text sections weren't
  marked up before.  I've marked them up for "or later" versions of
  the license (but not the "only" versions, since I'm leaving them
  alone for this commit).

* GPL-2.0+, GPL-2.0-or-later, LGPL-2.1+, LGPL-2.1-or-later: Added a
  missing paragraph for the "Also add information..." sentence after
  the recommended grant text.

* Interbase-1.0: Removed a typo'd "<" from the recommended grant text.
  I also standardized the LaTeX-ish '``' to '"' at the beginning of
  the new grant text.  This licence remains wildly inconsistent about
  quotation characters, but that was the only instance of '``'.

* LGPL-2.0+, LGPL-2.1+: The old recommended grant text did not include
  *the criticial* "any later version" wording.

* LGPL-2.0-or-later, LGPL-2.1-or-later: The old recommended grant text
  used "version 2 or any later version" instead of the FSF's
  recommended "any later version" phrasing.

* LPPL-1.1: Added some missing line-wrappings to the license text
  where we used to have mid-line % comment characters.

This commit does not touch BitTorrent-1.1, despite having an Exhibit A
with the recommended grant text.  I want to fix the referenced license
version from "1.0" -> "1.1", with "1.1" in our canonical text to
reduce future reader confusion.  But Gary felt that needed further
legal review [1,2], so I'm leaving that license alone for now.

This commit does not touch LPPL-1.3*, despite their including the
recommended grant text.  The <alt> matching for:

  This Current Maintainer of this work is <alt name="maintainer"
  match=".+">M. Y. Name</alt>.

is broken in spdx/tools 2.1.10 [3] (despite that pattern occuring in
the existing standardLicenseHeader), and I don't want to hold up this
branch.

[1]: https://github.com/spdx/license-list-XML/pull/581#discussion_r163964376
[2]: https://github.com/spdx/license-list-XML/pull/581#discussion_r163965036