schema/ListedLicense: Drop isDeprecated
deprecatedVersion should be set if and only if isDeprecated was true.
So instead of recording that information twice, tooling which consumes
this format should just check deprecatedVersion.  If deprecatedVersion
is non-empty, the license is deprecated.  If deprecatedVersion is
empty, the license is not deprecated.

This protects us from desynchronization like that fixed by cc1e05e8
(added deprecated version, 2017-12-28), because it is now impossible
to deprecate a license without recording the deprecated version.