LPPL-1.3*: Move standardLicenseHeader inside text
And add <alt> tags to cover the maintainance status.

The LPP-1.3a grant text has a "This Current Maintainer of this
work..." typo, which should be "*The* Current Maintainer...".  The
typo was fixed in 1.3c, but the upstream 1.3a text [1] still has
"This".  I've added an <alt> for this as well, with the upstream
"This" as canonical.

[1]: https://www.latex-project.org/lppl/lppl-1-3a.txt