Sendmail: Drop page-end 139848.1 from the canonical text
In the upstream PDF (linked with a crossRef to the archive.org URI)
this text is part of the footer on every page.  We've had it in this
repo since ef01296f (Automated, 2016-03-30, #212), but I imagine this
is a copy/paste issue dating back to the original SPDX version of this
license (sometime around 2015-09-16 [1]).

My personal preference would be to continue to match this via an
<alt>, but the list maintainers feel that is not necessary [2].

[1]: https://lists.spdx.org/pipermail/spdx-legal/2015-September/001497.html
     Subject: SPDX Legal call this Thursday
     Date: Wed Sep 16 00:33:04 UTC 2015
[2]: https://github.com/spdx/license-list-XML/pull/609#issuecomment-375380201