LLVM-exception: Replace optional dash-rules with alts
Use an alt instead of an optional because there's no legal
significance to the number of dashes (if any) used to set off the
header.

The alt is empty, because I see the dashes as header markup.  I'd much
rather they not be part of our canonical HTML, and instead would
rather use <h1> or other canonical HTML markup for "this is a header".
We don't have "this is a header" markup in our XML schema yet, but I'd
still rather not have these dashes be a part of the canonical HTML we
display under spdx.org/licenses/.