Nokia-Qt-exception-1.1: Fix doubled quote ("") typos
We've had these in this repository since the exception landed in
e3555218 (Adding exception: Nokia-Qt-exception-1.1, 2017-02-05, #354),
and they've been in license-list-data since the exception landed there
in spdx/license-list-data@cec28dc7 (Updated JSON exception files,
2016-04-16).  I haven't bothered with an <alt> tag because the
matching guidelines have [1]:

  5.1.3 Guideline: Quotes Any variation of quotations (single, double,
  curly, etc.) should be considered equivalent.

That doesn't explicitly say that you can add/remove quotes in the
general case.  And maybe that's intentional.  But if we treat the
whole unit as a quote (going from quad-quotes to double-quotes) then
this change should have no impact on matching.

In practice, I think it's unlikely that folks copied the exception
with the quad-quotes.  And the upstream linked from our XML is pinned
to a specific revision and only uses double-quotes.

[1]: https://spdx.org/spdx-license-list/matching-guidelines