Qt-LGPL-exception-1.1: Add new ID for Nokia-Qt-exception-1.1
And deprecate the old ID, which had two problems [1]:

* Qt no longer belongs to Nokia, and they've changed the title
  upstream to reflect that.
* Qt has another exception for GPL3, and the old ID didn't mention the
  target license.

This commit adds the new identifier and deprecates the old one in its
favor.

I've used Qt-LGPL-exception-1.1 instead of Kai's suggested
Qt-exception-LGPL-1.1 to match the {name}-{target}-exception-{version}
pattern suggested by the existing i2p-gpl-java-exception.

I've set an <alt> in the exception title so it will match the old
"Nokia " prefix, the current "The Qt Company " prefix, or no prefix at
all.  I've used "The Qt Company " prefix in the test file to match the
upstream text (the test is a copy/paste of the upstream text with
trailing whitespace removed).

[1]: https://lists.spdx.org/pipermail/spdx-legal/2018-March/002511.html
     Subject: New License/Exception Request: Qt-LGPL-exception-1.1
     Date: Fri, 23 Mar 2018 15:39:54 +0000
     Message-ID: <HE1PR0202MB2571F0B9F0305FA2150A428F98A80@HE1PR0202MB2571.eurprd02.prod.outlook.com>