*: Add <obsoletedBys> wrapping <obsoletedBy>
The Node tests were choking with:

  Entity: line 65: element element: Schemas parser error : Element
  '{http://www.w3.org/2001/XMLSchema}element': Invalid value for
  maxOccurs (must be 0 or 1).

which is mentioned in [1].  Despite unbounded being in the 1.0 spec,
maxOccurs was restricted to 1 inside <all> in XSD 1.0. The
restriction was lifted in 1.1 [2].  The backing libxml issue is [3].

This commit works around the limitation by adding a wrapping
<obsoletedBys> to mirror our existing <crossRefs> wrapping
<crossRef>.  In both cases, the wrapper adds no useful semantics, but
we need it to get the tests passing until we can find an XSD 1.1
validator.

[1]: https://github.com/albanm/node-libxml-xsd/issues/12
[2]: https://www.w3.org/TR/xmlschema11-1/#ch_models
[3]: https://bugzilla.gnome.org/show_bug.cgi?id=765936