MIT-0: Add the license used for some AWS samples
Previous discussion in [1,2].  The MIT-0 ID is from [1].

The "MIT No Attribution" title is from [1].  The "The MIT-Zero
License" title is from [3] (linked from [4]).  The empty-title case
explicitly supports instances like [5] which leave the title off.

The copyright line's optional "(c)" is for [3].  The (c) is also in
the vanilla MIT template.  The "(c)" is not part of our canonical
template though, because it's not in [6]

I have not included ". All Rights Reserved." in the copyright
template.  It is not in the vanilla MIT, [3], or [6].  [5] includes
it, possibly borrowing from BSD-2-Clause or similar (more on this in
85df23f80, BSD-2-Clause: Make "All rights reserved" optional,
2017-12-19, #527).  The copyright regexp is broad enough to cover this
text when it is included.

The test content is copy/pasted from [6].

[1]: https://lists.spdx.org/pipermail/spdx-legal/2018-March/002532.html
     Subject: New License/Exception Request: MIT No Attribution
     Date: Wed, 28 Mar 2018 17:44:35 +0000
     Message-ID: <d853b899f23048cbb16c9d6fe5b5a764@EX13D18UWC003.ant.amazon.com>
[2]: https://github.com/spdx/license-list-XML/issues/619
[3]: https://romanrm.net/mit-zero
[4]: https://github.com/aws/mit-0#see-also
[5]: https://github.com/awsdocs/aws-cloud9-user-guide/blob/master/LICENSE-SAMPLECODE
[6]: https://github.com/aws/mit-0#license-text