Add separate -RFN and -no-RFN identifiers for OFL licenses
As discussed in #724, OFL-1.0 and OFL-1.1 have options permit a licensor
to designate OFL-licensed fonts as having a "Reserved Font Name." If so
designated, then certain provisions of the license are active.

The existing SPDX entries for OFL-1.0 and OFL-1.1 do not distinguish
between these options, so current use of "OFL-1.0" or "OFL-1.1" cannot
be assumed to assert anything about whether a Reserved Font Name is,
or is not, present.

Additionally, using "WITH Reserved-Font-Name" as an exceptions list
entry does not seem appropriate, because the Reserved Font Name is not
really an "exception" -- rather, it imposes additional obligations if
it is present.

This commit adds new license list entries for:
*  OFL-1.0-RFN
*  OFL-1.0-no-RFN
*  OFL-1.1-RFN
*  OFL-1.1-no-RFN
to enable optionally designating whether a Reserved Font Name applies.
The existing OFL-1.0 and OFL-1.1 identifiers, without a suffix, continue
to have their presumed existing meaning that they should not be read
as asserting whether or not a Reserved Font Name is present, and this is
clarified in the Notes sections for those licenses.

Finally, because the license text is identical for these variants, they
have been added to the expected-warnings file to suppress the build
error to that effect.

Fixes #724

Signed-off-by: Steve Winslow <steve@swinslow.net>