Add GPL-3.0-linking-exception and GPL-3.0-linking-source-exception
This adds two variants of the recommended template exception from
the FSF's FAQ:
https://www.gnu.org/licenses/gpl-faq.en.html#GPLIncompatibleLibs

The only difference is that the "-source" variant incorporates the
second optional sentence from the FAQ template.

Note that in the identifier, "3.0" is intentionally included after
"GPL" rather than at the end of the identifier. That's because the
exception itself is not versioned, and this is meant to communicate
that the exceptions are intended for use with GPL-3.0.

Signed-off-by: Steve Winslow <steve@swinslow.net>