Deprecate BSD-2-Clause-NetBSD
This patch deprecates BSD-2-Clause-NetBSD in favor of the standard
BSD-2-Clause license, as discussed in #980.

For those who do choose to continue using it, this also fixes the
markup for the "This code is derived from..." line so that it does
not render on the license list as an incomplete sentence.

Signed-off-by: Steve Winslow <steve@swinslow.net>