Fix identifiers and add listVersionAdded
The only changes here are to swap the license identifiers as
discussed in #1043, instead of what was proposed in the initial
draft PR:

NIST-PD => NIST-PD-fallback
NIST-PD-old => NIST-PD

The diff shows up as much more extensive because Git is confused
about the file renames vs. deleting and recreating new files.

The only other change was to add listVersionAdded.

Signed-off-by: Steve Winslow <steve@swinslow.net>