Deprecate BSD-2-Clause-FreeBSD and add generic version
This commit adds BSD-2-Clause-Views, a generic version of
BSD-2-Clause-FreeBSD that templatizes the FreeBSD-specific
language. -Views is identical to standard BSD-2-Clause with the
addition of the "views and conclusions" final sentence.

It also deprecates BSD-2-Clause-FreeBSD, to avoid further
confusion, as developers from the FreeBSD community have indicated
that it was not intended to be part of that project's license for
source code.

Signed-off-by: Steve Winslow <steve@swinslow.net>