Rename Linux man-pages project license
Alejandro Colomar, Linux man-pages project co-maintainer, and I recently
discussed the name of this license and agreed that use of the term
"verbatim" is misleading, in that it could suggest to the reader that
the material so licensed withholds permission to distribute modified
versions.  We agreed upon a new name that captures the specificity of
this license to the project (the warranty disclaimer is worded such it
is ill-suited for literal reuse by other projects) and more accurately
suggests the set of granted permissions.

He furthermore recommended that I bring this proposed name change to the
SPDX team since this license is slated for inclusion in the next release
of the SPDX License List.

See
<https://lore.kernel.org/linux-man/8b513d06-3e9e-3d2f-6217-07bf2fa3765c@gmail.com/>
for supporting discussion.