Fix matching error to original license for CPAL-1.0
There is no space before the underscores in the original code for the line `The Original Code is___`.  This PR will allow matches for either no space or a space between the `is` and the underscores.

Note that once the next release of the LicenseListPublisher is used for checking, a match failure will occur if this PR is not merged.