Reformat for ease of use/skimming
Reformatted to ordered lists. Each list item has Yes/No checkboxes. The checkboxes are all at the same indent level, so the eye can skim down them to see the status at a glance.

This should make it easier both to fill in the form and also to review it.