Use relative Markdown links for stuff stored in this repo
There’s two different ways of viewing rendered versions of the Markdown
files in this repo. You can either

1. view them on GitHub or
2. clone the repo and open them using a program like Okular [1].

Before this change, option 2 was a little bit inconvenient. When you
clicked on many of the links in this repo’s Markdown files, it would
take you to GitHub instead of opening up the Markdown file locally.

[1]: <https://okular.kde.org/>