More directly draw attention to rest of list
Before this change, there was an awkward and slightly incorrect part of
the FAQ [1]. It was trying to draw attention to specific parts of the
License List Web page that typically get ignored [2].

This replaces that part of the FAQ with a more direct call out.
Hopefully, this will make people pay more attention to those parts of
the License List.

This change also removes an invalid link.

[1]: <https://github.com/spdx/license-list-XML/pull/1809#issue-1574869665>
[2]: <https://github.com/spdx/license-list-XML/pull/1809#issuecomment-1426582567>